+++ src/utils/sortObject.js
@@ -1,0 +1,17 @@
+const sortObject = obj => {
+   var sorted = {};
+   var str = [];
+   var key;
+   for (key in obj) {
+      if (obj.hasOwnProperty(key)) {
+         str.push(encodeURIComponent(key));
+      }
+   }
+   str.sort();
+   for (key = 0; key < str.length; key++) {
+      sorted[str[key]] = encodeURIComponent(obj[str[key]]).replace(/%20/g, "+");
+   }
+   return sorted;
+};
+
+module.exports = sortObject;
